+++ R/additive.R
@@ -146,6 +146,22 @@ additive <-
       fit = rlang::enquo(fit)
     )
 
+    if (is.null(family)) {
+      if (mode == "classification") {
+        args$family <- rlang::expr(
+          stats::binomial(
+            link = "logit"
+          )
+        )
+      } else {
+        args$family <- rlang::expr(
+          stats::gaussian(
+            link = "identity"
+          )
+        )
+      }
+    }
+
     parsnip::new_model_spec(
       "additive",
       args = args,